requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
arrow==1.0.3
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.23
discord.py==1.7.1
fuzzywuzzy==0.18.0
httplib2==0.19.1
idna==3.1
kiwisolver==1.3.1
matplotlib==3.4.1
mongoengine==0.23.1
more-itertools==8.8.0
multidict==5.1.0
numpy==1.20.1
Pillow==8.2.0
protobuf==3.15.8
psutil==5.8.0
pushover-nythepegasus==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.1
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
reactionmenu==1.0.7
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
sentry-sdk==1.0.0
six==1.15.0
speedtest-cli==2.1.3
typing-extensions==3.7.4.3
uptime==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
yarl==1.6.3
